pyproject.toml
[tool.poetry]
name = "blockchain-messanger"
version = "0.1.0"
description = ""
authors = ["Serchaev <german.serchaevk2.k2@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.26.0"}
pydantic = {extras = ["email"], version = "^2.5.3"}
pydantic-settings = "^2.1.0"
SQLAlchemy = {extras = ["asyncio"], version = "^2.0.25"}
asyncpg = "^0.29.0"
alembic = "^1.13.1"
fastapi-cache2 = {extras = ["redis"], version = "^0.2.1"}
aioredis = {extras = ["hiredis"], version = "^2.0.1"}
celery = "^5.3.6"
flower = "^2.0.1"
psycopg2-binary = "^2.9.9"
pysecurechain = "^0.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
black = "^23.12.1"
pytest-asyncio = "^0.23.5"
httpx = "^0.26.0"
pytest-cov = "^4.1.0"
isort = "^5.13.2"
flake8 = "^7.0.0"

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
